riscv: Add FPIOA and GPIO support for Kendryte K210
This patch adds the necessary configs and docs for FPIOA and GPIO support
on the K210.

Patch-cc: bmeng
Patch-cc: rchen
Signed-off-by: Sean Anderson <seanga2@gmail.com>
Series-to: uboot
Series-cc: trini
Series-cc: sjg
Series-version: 1
Cover-letter:
riscv: Add FPIOA and GPIO support for Kendryte K210
This patch series adds support for pinmuxing, gpios, and leds on the Kendyte
K210.

This patch series was previously part of
https://patchwork.ozlabs.org/project/uboot/list/?series=161576

This patch series depends on
https://patchwork.ozlabs.org/project/uboot/list/?series=162643
END